InfoView and HelpViews
Added info and help buttons to the navigation-bar..
Created a good-looking information view..
Created a custom BB-code parser to do custom formatting in UITextViews..
Photoshopped some navigation-bar icons and application icons, etc..

+++ MinPairs/MLPQBaseViewController.m
@@ -256,7 +256,8 @@
         
         
     }
-    int randomIndex=arc4random_uniform(filteredArr.count);
+    
+    NSUInteger randomIndex = arc4random_uniform((unsigned int)filteredArr.count);
     return [filteredArr objectAtIndex:randomIndex];
 }
 -(void) pushSequeOnStack:(NSNumber*)mode
